feat: implement class management features including fetching, joining, and creating classes with validation
